+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = 'Grid2Op a Series of LF Projects, LLC,\nFor website terms of use, tr
 author = 'Benjamin Donnot'
 
 # The full version, including alpha/beta/rc tags
-release = '1.12.3.dev0'
+release = '1.12.3'
 version = '1.12'